requirements.txt
openai>=1.30.0
PyYAML>=6.0
prompt_toolkit>=3.0.52
